requirements.txt
pytest~=7.3.1
selene~=2.0.0.rc2
allure-pytest~=2.13.2
